Moving database test class to another package and renaming variables
